+++ backend/backend/api/routes/room.py
@@ -1,7 +1,7 @@
 from functools import lru_cache
 from typing import Annotated
 
-from fastapi import APIRouter, Depends, status, HTTPException
+from fastapi import APIRouter, Depends, status
 
 from backend.common.room import Room
 from backend.common.member import RoomMember
@@ -9,24 +9,10 @@ from backend.common.member import RoomMember
 from backend.api.models.room import RoomId
 from backend.api.models.player import Player
 
-from backend.api.routes.player import get_other_by_id, get_player_by_id
+from backend.api.dependencies.room import rooms, get_room_by_id
+from backend.api.dependencies.player import get_player_by_id, get_other_by_id
 
 router = APIRouter(tags=["rooms"])
-
-
-@lru_cache
-def rooms() -> dict[str, Room]:
-    return {}
-
-
-def get_room_by_id(room_id: str,
-                rooms: Annotated[dict[str, str], Depends(rooms)]):
-    room = rooms.get(room_id)
-    if not room:
-        raise HTTPException(detail=f"room id {room_id} does not exist!",
-                            status_code=status.HTTP_404_NOT_FOUND)
-
-    return room
 
 
 @router.post("/rooms/{player_id}", status_code=status.HTTP_201_CREATED,
@@ -43,7 +29,7 @@ async def create_room(player: Annotated[Player, Depends(get_player_by_id)],
     return RoomId(room_id=room.room_id)
 
 
-@router.post("/enter_room/{room_id}/{player_id}",
+@router.post("/rooms/{room_id}/{player_id}",
              status_code=status.HTTP_202_ACCEPTED)
 async def enter_room(room: Annotated[Room, Depends(get_room_by_id)],
                      player: Annotated[Player, Depends(get_player_by_id)]):
@@ -59,14 +45,14 @@ async def enter_room(room: Annotated[Room, Depends(get_room_by_id)],
     return RoomMember
 
 
-@router.post("/room_players/{room_id}",
+@router.post("/rooms/{room_id}/players",
             status_code=status.HTTP_200_OK,
             response_model=list[RoomMember])
 async def get_room_players(room: Annotated[Room, Depends(get_room_by_id)]):
     return list(room.members.values())
 
 
-@router.put("/room_leader/{room_id}/{player_id}/{other_id}",
+@router.put("/room_leader/{room_id}/leader",
             status_code=status.HTTP_204_NO_CONTENT)
 async def update_room_leader(room: Annotated[Room, Depends(get_room_by_id)],
                              player: Annotated[Player, Depends(get_player_by_id)],
@@ -77,6 +63,6 @@ async def update_room_leader(room: Annotated[Room, Depends(get_room_by_id)],
 
 @router.delete("/rooms/{room_id}",
                status_code=status.HTTP_200_OK)
-async def delete_room(room: Annotated[Room,Depends(get_room_by_id)],
+def delete_room(room: Annotated[Room,Depends(get_room_by_id)],
                       rooms: Annotated[dict[str, Room], Depends(rooms)]):
         return rooms.pop(room.room_id)
